Renamed folders bug
